+++ src/lib.rs
@@ -4,26 +4,26 @@
 // enable const_generics if the binomial feature is used.
 #![cfg_attr(feature = "binomial", feature(const_generics))]
 
-// opt-in for binomial regression as it utilizes unstable features.
-#[cfg(feature = "binomial")]
-pub mod binomial;
 pub mod error;
 mod fit;
 mod glm;
-pub mod linear;
+mod irls;
 pub mod link;
-pub mod logistic;
 mod math;
 pub mod model;
-pub mod poisson;
+mod num;
 mod regularization;
-pub mod standardize;
+mod response;
+mod standardize;
 mod utility;
 
 // Import some common names into the top-level namespace
 #[cfg(feature = "binomial")]
-pub use binomial::Binomial;
+pub use response::binomial::Binomial;
 pub use {
-    linear::Linear, logistic::Logistic, model::ModelBuilder, poisson::Poisson,
+    model::ModelBuilder,
+    response::{
+    linear::Linear, logistic::Logistic, poisson::Poisson,
+    },
     standardize::standardize,
 };
