+++ src/num.rs
@@ -1,0 +1,2 @@
+//! numerical trait constraints
+
