+++ src/response/poisson.rs
@@ -74,7 +74,8 @@ pub mod link {
 
 #[cfg(test)]
 mod tests {
-    use crate::{error::RegressionResult, model::ModelBuilder, poisson::Poisson};
+    use super::*;
+    use crate::{error::RegressionResult, model::ModelBuilder};
     use approx::assert_abs_diff_eq;
     use ndarray::{array, Array1};
